firestoreEmulatorProxyを切り出し、さらにvite/pluginを整理
