+++ vite/plugin/bundleBuddy.ts
